requirements.txt
Django
djangorestframework
django-cors-headers
drf-spectacular
python-dotenv
PyMySQL
chromadb
selenium
ollama
sentence-transformers
requests
